simplify some functions
